feat(cli): track early start instances and add StartInfo
Add logic to detect and track instances that are in the process of
starting but not yet fully initialized, using status files with a
short validity window. Introduce a StartInfo struct to provide
information about the starting testnet instance, including PID and
logfile. Prevents race conditions when allocating instances and
improves reporting of start status.